fix: retrieve candidate sessions without id
