+++ src/app/api/sessions/candidate/[id]/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest } from 'next/server';
-import { withAuthAndDB, errorResponse, successResponse } from '@/lib/api/error-handler';
-import Session from '@/lib/models/Session';
+import { withAuthAndDB, errorResponse } from '@/lib/api/error-handler';
 import type { Session as AuthSession } from 'next-auth';
+import { getCandidateSessions } from '../route';
 
 /**
  * GET /api/sessions/candidate/[id]
@@ -25,79 +25,7 @@ export const GET = withAuthAndDB(async (
     return errorResponse('Unauthorized - not your sessions', 403);
   }
 
-  const now = new Date();
-
-  // Fetch upcoming sessions (only confirmed and still in the future)
-  const upcomingSessions = await Session.find({
-    candidateId,
-    scheduledAt: { $gte: now },
-    status: 'confirmed'
-  })
-  .populate('professionalId', 'name title company profileImageUrl')
-  .sort({ scheduledAt: 1 })
-  .lean();
-
-  // Fetch completed sessions 
-  const completedSessions = await Session.find({
-    candidateId,
-    $or: [
-      { status: 'completed' },
-      { 
-        status: 'confirmed',
-        scheduledAt: { $lt: now }
-      }
-    ]
-  })
-  .populate('professionalId', 'name title company profileImageUrl')
-  .sort({ scheduledAt: -1 })
-  .limit(50) // Limit to recent 50 completed sessions
-  .lean();
-
-  // Get pending sessions (requested but not yet confirmed)
-  const pendingSessions = await Session.find({
-    candidateId,
-    status: 'requested'
-  })
-  .populate('professionalId', 'name title company profileImageUrl')
-  .sort({ createdAt: 1 })
-  .lean();
-
-  const monthlySpending = await calculateMonthlySpending(candidateId);
-
-  return successResponse({
-    upcoming: upcomingSessions,
-    completed: completedSessions,
-    pending: pendingSessions,
-    stats: {
-      totalUpcoming: upcomingSessions.length,
-      totalCompleted: completedSessions.length,
-      totalPending: pendingSessions.length,
-      totalSpentThisMonth: monthlySpending
-    }
-  });
+  return getCandidateSessions(session);
 
 }, { requireRole: 'candidate' });
 
-/**
- * Calculate total spending for current month
- */
-async function calculateMonthlySpending(candidateId: string): Promise<number> {
-  const now = new Date();
-  const startOfMonth = new Date(now.getFullYear(), now.getMonth(), 1);
-  const endOfMonth = new Date(now.getFullYear(), now.getMonth() + 1, 0);
-
-  const paidSessions = await Session.find({
-    candidateId,
-    status: { $in: ['confirmed', 'completed'] },
-    paidAt: {
-      $gte: startOfMonth,
-      $lte: endOfMonth
-    }
-  }).select('rateCents');
-
-  const totalCents = paidSessions.reduce((sum, session) => {
-    return sum + session.rateCents;
-  }, 0);
-
-  return Math.round(totalCents / 100); // Return in dollars
-}
